requirements.txt
asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.4.0
django-filter==24.3
django-unfold==0.40.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
PyJWT==2.9.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
